[shell][mac] Fix Ninja builds
Ninja builds are the fastest, so it's useful to have it supported.

The cmake code currently has some logic that must be handled differently for Xcode vs. Makefiles, but it's more correct for it to be Xcode vs. anything else. It worked for "Xcode" and "Unix Makefiles", but now also works for "Ninja" (and maybe others?).

Test steps:
$ mkdir build-ninja
$ cd build-ninja
$ cmake .. -G Ninja -DIGL_WITH_VULKAN=OFF
$ ninja