Load files from bundle path where applicable (iOS & Mac)
Summary: Add an API fullPath to extract the full path of a file.  We first look for a resource bundle path.  If that fails, we fall back to the regular file path.  This enables us to load assets from resource bundles on platforms that support resource bundles (e.g. mobile).

Reviewed By: corporateshark

Differential Revision: D47473530

fbshipit-source-id: 926c03f44a2d33977a5f2a269542a72bee3b57d7