IGL: Separately specify face and layer for render passes
Summary:
This diff updates RenderPass attachment descriptors to have separate parameters for face and layer. This makes usage unambiguous.

On OpenGL, this diff also cleans up how textures and renderbuffers are attached to a framebuffer in OpenGL to accomodate this change. All attachment logic is consolidated in TextureBufferBase / TextureTarget. Attaching color, depth and stencil all reuses the same function. All possible params are encapsulated in a new Texture::AttachmentParams struct. Framebuffer is updated to populate the correct params depending on whether reading or writing.

Differential Revision: D48378465

fbshipit-source-id: 297cb9827f52dc7b417dde071750a7dcd25a9cd8