igl | vulkan | Remove unused flag awaitingDeletion
Summary:
Our bindless descriptor set was created with `VK_DESCRIPTOR_BINDING_PARTIALLY_BOUND_BIT` which indicates that descriptors in this binding that are not dynamically used need not contain valid descriptors at the time the descriptors are consumed. A descriptor is dynamically used if any shader invocation executes an instruction that performs any memory access using the descriptor. If a descriptor is not dynamically used, any resource referenced by the descriptor is not considered to be referenced during command execution.
https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VkDescriptorBindingFlagBits.html

Reviewed By: pixelperfect3

Differential Revision: D48543715

fbshipit-source-id: c0f9bd423c101d3394eae8d02056bf06af2e2ecf