Give Vulkan a verbose error when the SDK isnt installed
Summary:
I was asked to test the Granite samples with Vulkan backend.

It died on startup with just an assertion because I didn't have the Vulkan SDK installed

It took me a few minutes to diagnose the issue. Perhaps it is best to have an explicit message?

Also I added all the casts that linter recommended.

Reviewed By: MichaelTay, corporateshark

Differential Revision: D48621596

fbshipit-source-id: 738516a381e7c76614f4e5857ab6f773ee38da10