IGL: Add withNumMipLevels helper and fix atMipLevel
Summary: This diff fixes a bug with the atMipLevel range helper and adds a new withNumMipLevels helper. The bug with atMipLevel was that if the new mip level was the same as the current one, the returned range had the same number of mip levels as the input range, when it should have had numMipLevels = 1.

Reviewed By: rokuz

Differential Revision: D48737529

fbshipit-source-id: 3fe5a5f321b211f7ebec46ce4623a0ca966e975a