IGL: Clean up repacking of Vulkan texture data
Summary:
Vulkan requires texture data that is uploaded or downloaded to have rows that are aligned to a multiple of a texel block. IGL supports an arbitrary byte length for rows of texture data. This requires repacking data when an upload source buffer or download destination do not have a row byte length that is a multiple of a texel block size.

This diff moves the current repacking functionality from Texture::upload to a helper method on VulkanStagingDevice. The new helper can handle both upload and download.

The upload code (imageData2D / imageData3D) and the download code (getImageData2D) in VulkanStagingDevice are updated to repack data when needed.

For texture uploading, this is a slight optimization in that data is only repacked if the row length doesn't align with a texel block size. That is, row lengths that are larger than necessary but ARE aligned with texel block size are not repacked.

Later diffs will increase the efficiency of upload to repack all texture data provided to upload at once.

Reviewed By: syeh1

Differential Revision: D48756197

fbshipit-source-id: 0ca20531a8fcc3c50fb63c2c5ce60a915f73e173