Call IGL method to flush Metal texture cache
Summary: This diff updates two places were accessing the texture cache held in an IGL Metal PlatformDevice and attempting to flush it without checking if it is null first. The update is to call the existing method on the PlatformDevice that does the same thing but includes the null check.

Reviewed By: syeh1

Differential Revision: D49151276

fbshipit-source-id: b12a846d7cc917affc8ff06b4e39629437b5b12f