IGL: Add load helpers to texture_loader
Summary: This diff adds new load and loadToExternalMemory helper methods to texture_loader. These are intended to support use cases where the texture data is not directly in a form usable by the GPU (e.g., png, jpeg). A new memorySizeInBytes is also added to help allocate memory to load the data to. A new supporting class, IData, is also introduced to hold data when the class itself must allocate the data.

Reviewed By: MichaelTay

Differential Revision: D49135958

fbshipit-source-id: 59d050e2a125b6f9c6b138ca765bebeb4abed3cd