Do not fail if a uniform location is updated more than once
Summary:
Now that we're migrating to uniform blocks (simulated with uniform struct instances in legacy OpenGL), there are cases where the same uniform location is updated twice.

As an example, in particle systems (ParticleSystems.cpp), we have the following sequence of logic
```
material -> clearUniform(X);
// submit all the uniforms common to all groups
material -> submit();

for(each particle group){
  compute X for this group;
  material -> setUniform(X);
  material -> submit();
}
```
Here, the clearUniform is a no-op, because it is a member in a block. So as a result we submit X multiple times.

For now, this diff disables the check for duplicate submissions to unblock XOR migration. We'll address the duplicate submission issue in the future.

Reviewed By: syeh1

Differential Revision: D49243113

fbshipit-source-id: 259be87dcd2bff37456ce0f68222a2798c5219f2