Always use GLSL ES 100 shaders in swift shaders
Summary:
OpenGL ES 3 on rainbow has a bug, which originates from a bug in swiftshaders.

the issue is that swiftshaders is returning -1 when we use
```
glGetActiveUniformsiv(,..,GL_UNIFORM_OFFSET, ...)
```
to query the offset of a member in a uniform buffer. We believe this is a swift shaders bug, so we work around this by always using GLSL es 100 shaders on rainbow

Reviewed By: dmannemela, syeh1

Differential Revision: D49329247

fbshipit-source-id: 54a8245704d932e7984e7521bef9d28367d1e416