Revert D49329247: Multisect successfully blamed "D49329247: [IGL] Always use GLSL ES 100 shaders in swift shaders" for test or build failures
Summary:
This diff is reverting D49329247
D49329247: [IGL] Always use GLSL ES 100 shaders in swift shaders by AmesingFlank has been identified to be causing the following test or build failures:

Tests affected:
- [cameracore/rainbow/snapshot_tests:gltf_analytical_spheres2 - Main/RainbowSnapshotTest.Generated/gltf_analytical_spheres2_gles2_hermes_vEffect_etc](https://www.internalfb.com/intern/test/562950033707889/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3059442
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: AmesingFlank

Differential Revision: D49344838

fbshipit-source-id: 79721d37497ce5afa0cb3184ce13b2ec996e8b55