IGL: Tidy up API logging of strings
Summary: This diff tidies up API logging for strings. When logging shader source, source is logged one line at a time to ensure larger programs are logged in full on Android. For other string return values, falling back to std::strlen when length is not provided as the spec guarantees a null terminator.

Reviewed By: corporateshark

Differential Revision: D47272414

fbshipit-source-id: 302d204b80c6e6e87eed9d9576bbac0291cfbbd0