IGL: Add Invalid backend type
Summary:
This diff adds a new backend type: Invalid. This will be used to create a safe, sentinel device that returns Invalid as its backend type, which should prevent any code that validate the backend of a device it works with from using it.

Relevant code that is checking all possible backends is updated to reflect this change.

Reviewed By: corporateshark

Differential Revision: D49359044

fbshipit-source-id: 84115bede9e59e695e814909705523c215e30e1e