Add IContext::createShareContext
Summary:
State of "shared contexts" before this diff:
* The only way to create a shared context is via the platform specific headers in igl/opengl/<platform>/. If that's not enough friction, the macos header is ObjC++ and can't be used from C++ code.
* Of all platforms, only `egl` and `macos` provide API for creating shared contexts.

This new API allows us to create a shared context from an IContext instance. This new generic method might not allow full customization across all implementations since that sometimes requires platform specific objects/information, so the assumptions here are (1) the new context is part of the same sharegroup, and (2) it's as similar to the existing context as possible.

Note #1: this diff does not implement the new API for all platforms because I don't have the means to test them all. IMO, it's fine for people to have to implement it as needed, since they'll clearly be able to test it.

Note #2: don't be fooled by `IDevice::getSharedContext`, it has nothing to do with multiple contexts in the same sharegroup.

Differential Revision: D49297661

fbshipit-source-id: cfc39b10d9310df9a8e854ae31e852c1e895720d