Add dynamically sized staging buffer
Summary:
This change modifies the `VulkanStagingDevice` to use a staging buffer that grows in size dynamically, according to demand.

At first, we don't allocate a staging buffer. One is allocated only when needed, ie, when uploading an image or data. On devices with shared heaps, a staging buffer is only allocated when uploading images.

The size doesn't decrease.

Reviewed By: corporateshark

Differential Revision: D49329489

fbshipit-source-id: c40fd45c11ce855ad29805277f986215a1330fae