IGL: Allow color+depth to be updated on Framebuffer via SurfaceTextures
Summary:
This diff updates the updateDrawable method on IFramebuffer to take two optional parameters to allow the depth and stencil textures to be updated.

Additional, the return is changed to void as no callers are doing anything with the return value.

For OpenGL, the ability to detach stencil and depth textures is added, and the ability to detach color renderbuffers is fixed.

Reviewed By: tgoulart, corporateshark

Differential Revision: D49401437

fbshipit-source-id: 2bedf625f6287de48aeba98e16bad14cd55eb30c