Exporting Android native hardware buffers out of IGL (take 2)
Summary:
This diff exposes an Android shared CPU+GPU memory buffer on EGL.

This is a second submission of D49734694 which broke the build on the Github Android version, so was reverted in D49760636.

This second pass adds a change to CMAKE to include the missing cpp file from the Github build.

Reviewed By: EricGriffith

Differential Revision: D49775587

fbshipit-source-id: c29ebca35dd0e4472d9a02a7fc084fc22c20b6e7