IGL: Remove helper functions from TextureFormat.h
Summary:
This diff removes two helper functions from TextureFormat.h. Each is only used in one place.

isTextureFormatBGR is inaccurately named as it only checks for two of the five supported BGR formats. It is be replaced with its implementation.

isTextureFormatsRGB is also a less functional version of checking isSRGB() on the format's properties so this call is replaced with using the texture format properties.

Reviewed By: KyleFung

Differential Revision: D49699222

fbshipit-source-id: af8811b765b999059621df3b5bccf7c1cf6c746e