Flip image vertically before saving framebuffer to PNG
Summary:
I believe in all of the current use cases we want to actually flip the image vertically when writing framebuffer contents as a screenshot, otherwise the screenshot that is saved is upside down, which potentially complicates troubleshooting.

So this diff does just that (tell stbi to do the flip, which also appears to be a common practice). It also updates the corresponding snapshots for the currently active tests.

One example before/after, this is for XorScreenshotTest_GLTF.png:
**Before:**
{F1109603554}

**After:**
{F1109603280}

So the screenshots are looking correct now.

The rationale for this is that for RN screenshot tests we expect to have a significant amount of test cases, which, among other things, will be compared vs reference images on existing platforms (e.g. VR or iOS). Having snapshots look correctly will simplify reasoning when troubleshooting.

Reviewed By: corporateshark

Differential Revision: D49868910

fbshipit-source-id: 67a15a2d214b2f1f88e5b84ae475ff870253dfbe