Revert D50436124: Multisect successfully blamed "D50436125: [IGL] Simplify `ShaderUniforms::setFloat3x3`" for test or build failures
Summary:
This diff is reverting D50436124
Depends on D50483433
D50436125: [IGL] Simplify `ShaderUniforms::setFloat3x3` by AmesingFlank has been identified to be causing the following test or build failures:

Tests affected:
- [//fbandroid/instrumentation_tests/com/facebook/msqrd:block_inputs_every_type_via_patches-64bit - testDefault[base_gles2_hermes_vEffect_uncompressed] (com.facebook.msqrd.unified_runner.AREngineScreenshotTest)](https://www.internalfb.com/intern/test/281475070193663/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3348209
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: syeh1

Differential Revision: D50483445

fbshipit-source-id: e265ad75ffcb9bdc274b1ce3ebd4081743d50b2f