Make sure samples' session is released before other global destructors
Summary:
This enforces release of the Session object before exiting from the main app function in shell.

This guarantees that whatever graphics resources are held in the Session object, they are released before the call of the other global destructors, one example of the latter being resource trackers.

Since the order of the global destructors' calls is not guaranteed, this may cause a situation when a leak tracker is triggered before the resources were released, triggering a false positive assert on exit.

Reviewed By: MichaelTay

Differential Revision: D50603617

fbshipit-source-id: 33ebe765d3e35e07760732dbdb988e7fd877455a