IGL: Fix framebuffer binding
Summary:
This diff fixes an issue where we are always binding the framebuffer to GL_FRAMEBUFFER for writes, even when we could bind it to GL_DRAW_FRAMEBUFFER. This is causing issues when we try to read data from the framebuffer via GL_READ_FRAMEBUFFER with texture_loader. This may have also been the cause of some recent MSAA related issues.

texture_loader is also updated to bind the framebuffer for read.

Reviewed By: nlutsenko, rameshviswanathan

Differential Revision: D49943436

fbshipit-source-id: d83210293007c8dd34467c531bdaad1ba9efb4ea