IGLU: Fix issues with texture accessor
Summary: This diff fixes an issue with IGLU texture_accessor: when using the async path, the code is not handling the case where the data is available in the first call to getRequestStatus. In the current code, if this happens, the data is never actually read from the PBO. This diff adds a new flag to track whether the data has been read or not, and ensure that the data is always read before returning from getBytes.

Reviewed By: rameshviswanathan

Differential Revision: D49943435

fbshipit-source-id: 10b123f214550bc930ac673a8fcfeeb26999c352