Do not override UBOs with pre-existing non-zero binding points (adressing Adreno bug)
Summary:
# This stack
This is for addressing a bug in Adreno's OpenGL driver. We noticed that, in certain old Adreno devices, if we specify the binding point of a block in the shader, and later tries to override it via `glUniformBlockBinding()` to explicitly override the binding points, then the driver won't like it, and might crash druing `glDrawElements()`.

Currently, for OpenGL, we rely on `RenderPipelineDesc::uniformBlockNames` to override binding points. This stack aims to get rid of this mechanism completely. Instead, in OpenGL, for both ARE and XOR, we will look up the proper binding point from reflection, and bind the buffer at that point. (In XOR, for non-OpenGL backends, we will still bind using the binding points specified by the Semantic).

# this diff
Do not override UBOs with pre-existing non-zero binding points (addressing Adreno bug)

Reviewed By: MichaelTay

Differential Revision: D50950795

fbshipit-source-id: b1cd62a1448a9c5ec8226b40fc5e89fac90073e9