Return `bindingIndex` instead of `blockIndex` for `getIndexByName()` when name is a uniform block name
Summary:
# This stack
This is for addressing a bug in Adreno's OpenGL driver. We noticed that, in certain old Adreno devices, if we specify the binding point of a block in the shader, and later tries to override it via `glUniformBlockBinding()` to explicitly override the binding points, then the driver won't like it, and might crash druing `glDrawElements()`.

Currently, for OpenGL, we rely on `RenderPipelineDesc::uniformBlockNames` to override binding points. This stack aims to get rid of this mechanism completely. Instead, in OpenGL, for both ARE and XOR, we will look up the proper binding point from reflection, and bind the buffer at that point. (In XOR, for non-OpenGL backends, we will still bind using the binding points specified by the Semantic).

# this diff
This diff is preparation work:
Return `bindingIndex` instead of `blockIndex` for `getIndexByName()` when name is a uniform block name. In OpenGL, each block is associated with two integers: a bindingIndex and a blockIndex, the bindingIndex is the one we're interested in.

Reviewed By: syeh1

Differential Revision: D50950796

fbshipit-source-id: 6667564984cf0fa4f9167f5cd0bf7ad3f744ea20