renderSessions comply to IGL backend selection once again
Summary: The shell framework complies to the [backend selection](https://www.internalfb.com/intern/wiki/IGL/Developer_Guide/#backend-selection) capability in IGL. Although individual shell apps are not forced to comply to it, the embedded `renderSessions` should since a lot of other targets depend on it, whatever the reason is. TinyMesh was the only app not complying.

Reviewed By: corporateshark

Differential Revision: D52065009

fbshipit-source-id: 8c441762158c31db5276fcbf5caf042c6e475770