Remove Previous Work Around
Summary:
Trying D51486102 again after check that I don't hit the assert on Quest 2.

This ASSERT is meant to catch issues elsewhere in the rendering stack, e.g. when someone else handed us a Vulkan surface that they rendered to without using the Engine (or IGL).

If this diff trips again, it's better to figure out the root cause and fix. However, when pressed, it's safe to back this out to buy time for the actual fix.

Reviewed By: corporateshark

Differential Revision: D51633218

fbshipit-source-id: 4fec3d197ac7d756bed6a127f50c2c9157d91d1f