Expose Entitlements.plist file for apps to sign with
Summary:
The `get-task-allow` entitlement is required to use the "Debug Memory Graph" tool within Xcode. I wasn't able to figure out how to integrate this entitlements file with shell's macosx_binary, but I can at least sign it manually after the app is built with [this command](https://stackoverflow.com/a/77269689).

Would be great to eventually get this fully integrated, if anyone knows how to deal with code signing + entitlements in buck.

Differential Revision: D52684052

fbshipit-source-id: e8abcc5dfde38327102a30789034743a327c1d26