Fix cmake shell builds not installing successfully on an iOS device
Build was successful, but installing failed with "Failure Reason: Failed to get the localized bundle name for the app to be installed. Recovery Suggestion: Ensure that your bundle's Info.plist contains a value for at least one of the following keys: CFBundleName or CFBundleDisplayName."

In this commit, I'm setting the CFBundleName and also using the shell session name as the identifier, so multiple shell apps can be installed on the same device without collision.

Test steps:
1. $ mkdir build & cd build
2. $ cmake .. -G Xcode -DCMAKE_TOOLCHAIN_FILE=../third-party/igl/third-party/deps/src/ios-cmake/ios.toolchain.cmake -DDEPLOYMENT_TARGET=14.0 -DPLATFORM=OS64 -DENABLE_BITCODE=OFF
3. Open and build any shell apps in the generated Xcode project. You may have to manually set the Development Team.
4. Verify the apps now run as expected.