WASM compilation fixes (#64)
Summary:
# WASM compilation fixes
## Description of the issue
This PR intend to solve the issue https://github.com/facebook/igl/issues/57. Following the documentation in the Readme to [build ](https://github.com/facebook/igl?tab=readme-ov-file#build) the WebAssembly artifacts:

```
cd build
emcmake cmake .. -G Ninja
cmake --build .
```

the build process fails as described in the linked issue https://github.com/facebook/igl/issues/57. I tested in both linux (Ubuntu 22.04) and Windows 11 platforms, and in both cases i have the same result.
with little changes i was able to build the **igl** lib and consequently the two examples in the wasm folder.

## List of changes

- In the main `CMakeLists.txt` file exclude **IGLU** submodule from compilation for EMSCRIPTEN build.
- In `igl/CMakeLists.txt` avoid linking of the **fmt** library for Emscripten.

part of the code of IGLU depend on ktx and that was failing the build process, not sure if in the past was possible to build the IGLU submodule with WASM i'm waiting a feedback from the maintainers in this sense.
I excluded **fmt** because reading in the code <fmt/core.h> header is included in `vulkan/Common.h` and other related vulkan code, as i know vulkan can not be compiled with Emscripten so i assumed that can be excluded.

Pull Request resolved: https://github.com/facebook/igl/pull/64

Reviewed By: EricGriffith

Differential Revision: D52848407

Pulled By: corporateshark

fbshipit-source-id: 3adda4dcb0e0a07ee576ea4a26afce16415b3b17