Rebrand toConstChar to c_str
Summary:
We intend to use **NameHandle** as replacement for **std::string** in cases it is used to frequently look or compare identifiers.
Current process of that migration is painful as we need to keep renaming **c_str** with **toConstChar**. This is pointless.

This change only inlines **toConstChar**, after refactoring code it will be removed.

Reviewed By: EricGriffith

Differential Revision: D52916755

fbshipit-source-id: 8463cce3ae9de8f1654679e35561a560cbb66ab2