[igl][shell][mac] Don't crash when tearing down the view controller
This is a feature that's only on mac shell builds. From the app's toolbar, click Debug -> Tear down ViewController. It does as the name suggests, and it's useful to debug memory leaks in Xcode.

Apparently, there's a known bug in MTKView where it's never removed from its internal DisplayLink (see https://openradar.appspot.com/23977735). This new check at least prevents it from crashing when trying to render with a null render session.