Update GPU feature support to new Metal API
Summary:
This diff simplifies things for the next one in the stack, which was the motivation behind this change, to fix an issue on iOS simulator (next diff).

Shift GPU feature capability checks to be based on GPU family as opposed to the deprecated featureSet Metal API.  This also simplifies the overall logic, since we don't need to deal with GPU family and their minor versions.  All of the feature logic was based on the GPU family anyway, and the new API forces us in this direction anyway.  This also unifies some of the logic between iOS and Mac.

Reviewed By: EricGriffith, KyleFung, corporateshark

Differential Revision: D54246362

fbshipit-source-id: 6243ba917b09ee8c2ef1dafcd840d965864e2c0a