Depth sample comparison was available as of Apple GPU family 2
Summary: Fix GPU capability check for depth sampler comparison.  This feature was available as of Apple GPU family 2, but we had it at GPU family 3+.  This fixes an issue where objects were rendering fully in shadow on iOS simulator, since the iOS simulator runs on a Mac with a desktop GPU.  My desktop has a GPU family 2, which is capable of depth sampler comparison, but we were saying it wasn't capable.

Reviewed By: KyleFung, corporateshark

Differential Revision: D54246364

fbshipit-source-id: 65fc790e2607170a9ae3c820e7619ecca1f31df4