Remove enum for the latest Apple GPU family 9 (A17, M3)
Summary:
We had to remove the enum MTLGPUFamilyApple9 (A17, M3) as it's breaking github builds of IGL.  It appears MTLDevice.h header in Xcode 14 and earlier didn't have this enum, as it was added in Xcode 15 versions of the same headers.
We can add it at a later point when the enum is more prominent.

Reviewed By: EricGriffith, KyleFung, corporateshark

Differential Revision: D54368037

fbshipit-source-id: 45e28cdbe74c0fc68ee132011cb0a46cc01dbf3a