Store unique ptr to VulkanImageView in VulkanTexture.
Summary:
This diff makes changes to the IGL Vulkan codebase to store a unique pointer to VulkanImageView in VulkanTexture. This change establishes VulkanTexture as the clear and sole owner of a VulkanImageView.

The changes include updating the code in VulkanTexture.cpp, ShellVulkanTextureRenderTarget.cpp, and IGLSurfaceTextureVulkan.cpp to use a unique pointer to VulkanImageView instead of a shared pointer. Additionally, changes were made to VulkanImage.cpp to create a unique pointer to VulkanImageView instead of a shared pointer.

Reviewed By: corporateshark

Differential Revision: D55260742

fbshipit-source-id: 32f79d009e448d0e1af2a0a0cd4b002ea1ef14a6