Store unique ptr to VulkanTextures in VulkanSwapchain.
Summary: This diff is part of a larger effort to improve the performance of IGL's Vulkan backend. The changes involve storing unique pointers to VulkanTextures in the VulkanSwapchain class, which allows for more efficient memory management. The code changes in each file of the diff include modifying the VulkanSwapchain.h and VulkanSwapchain.cpp files to use unique pointers instead of shared pointers, and modifying the PlatformDevice.cpp file to use the unique pointer instead of a shared pointer.

Reviewed By: corporateshark

Differential Revision: D55262401

fbshipit-source-id: 01499433f729af7f3ba2817fc668252432cb907e