igl | ITrackedResource : Address ubsan issue 2/2
Summary:
'vptr' sanitizer mode checks the use of an object whose vptr indicates that it is of the wrong dynamic type, or that its lifetime has not begun or has ended.

We have a problem in ITrackedResource, where ubsan is balking when we  downcast ITrackedResource, whose destructor is called after the object is partially destroyed already.

This diff turns the 'vptr' sanitizer off for the static_cast

Reviewed By: syeh1, tgoulart

Differential Revision: D55626887

fbshipit-source-id: 85309063c0d6149dbbcfc6805c4bb9433cfb3d91