igl | ITrackedResource : pass ptr instead of reference 1/2
Summary:
**Stack** :
'vptr' sanitizer mode checks the use of an object whose vptr indicates that it is of the wrong dynamic type, or that its lifetime has not begun or has ended.

We have a problem in ITrackedResource, where ubsan is balking when we  downcast ITrackedResource, whose destructor is called after the object is partially destroyed already.

**Diff**:
This diff switches to using pointers to pass arguments to IResourceTracker, to make it explicit that we only need the  ptrs and do not actually need to do deference.

Reviewed By: syeh1, MizFB

Differential Revision: D55627620

fbshipit-source-id: abb6935661e5f13a4047009d2989ee287f4e7365