Add In-Development API
Summary:
This will allow IGL clients to experiment and to suss out perf
and stability issues with more risky IGL changes, reverting to a
safer path at run time when necessary.

Note that it is up to the client of IGL to define what the bitfields
mean and what actions to take in their own copy of IGL as the feature
is in development.

No in-development paths will be upstreamed or accepted by public IGL repo.

Reviewed By: tgoulart, corporateshark

Differential Revision: D55597260

fbshipit-source-id: 9bdd5c5ef40acf31f84e3af10d23e9c408998c6e