Storing Buffer::buffers_ as array from vector.
Summary: The change is to store the buffers as an array of unique pointers to VulkanBuffer objects instead of using a vector. This is done to improve performance and reduce memory usage. The change also adds a new member variable to keep track of the number of buffers in the array.

Reviewed By: corporateshark

Differential Revision: D55577857

fbshipit-source-id: 38094fcde6e479299abc3628c0be9c9f078e85ef