Making VulkasnStagingBuffer::getBufferSubData take in const VulkanBuffer.
Summary: This diff makes a change to the `VulkanStagingDevice` class in the IGL library. Specifically, it modifies the `getBufferSubData` function to take in a `const VulkanBuffer` instead of a non-const reference. This change is necessary to ensure that the function does not modify the buffer object passed in as an argument.

Reviewed By: EricGriffith

Differential Revision: D55577044

fbshipit-source-id: 8496d719257907e576fb89fa458c39732a9549c7