Adding VulkanImageView::destroy function to destroy vk image view.
Summary: This diff adds a new function to the `VulkanImageView` class called `destroy` which is responsible for destroying the Vulkan image view. The `destroy` function is called in the destructor of the `VulkanImageView` class and uses the `vkDestroyImageView` function to destroy the image view. The `destroy` function also sets the `vkImageView_` to `VK_NULL_HANDLE` and the `ctx_` to `nullptr` to ensure that the image view is not accessed after it has been destroyed.

Reviewed By: corporateshark

Differential Revision: D55636136

fbshipit-source-id: 9f40a25fda7d2397e7335ed2deba25da93179d2b