Making VulkanStagingDevice::imageData image input const.
Summary: This diff makes a change to the VulkanStagingDevice class in the IGL library. The imageData function now takes a const reference to the VulkanImage object instead of a non-const reference. This change is made to ensure that the function does not modify the image object, which is a const object.

Reviewed By: corporateshark

Differential Revision: D55607100

fbshipit-source-id: 4a6e1612fdd04662cccfd242c83c2d29aa5a4f68