Storing VulkanImageView object in VulkanTexture instead of unique ptr reference.
Summary: This diff makes changes to the IGL Vulkan codebase to store a VulkanImageView object in a VulkanTexture instead of using a unique_ptr reference. This change is made to improve performance and reduce memory usage. The changes include modifying the VulkanTexture class to store a VulkanImageView object instead of a unique_ptr reference, modifying the VulkanImage class to create a VulkanImageView object instead of returning a unique_ptr reference, and modifying the VulkanContext class to use the VulkanImageView object instead of the unique_ptr reference.

Reviewed By: corporateshark

Differential Revision: D55600052

fbshipit-source-id: f932b20ed7c98fe13b487ba4095032e82233d216