Adding VulkanImage::destroy function to destroy vk image.
Summary: This diff adds a new function to the `VulkanImage` class called `destroy()` which is responsible for destroying the Vulkan image. The function is called in the destructor of the `VulkanImage` class and checks if the image has been externally managed before destroying it. The function also sets the `ctx_` and `vkImage_` variables to null to ensure that they are not used after the image has been destroyed.

Reviewed By: corporateshark

Differential Revision: D55648133

fbshipit-source-id: b9ca935da64f1fbc73cc6c7f3ae7770cfebb650d