Using multiple buffers for VulkanStagingDevice to improve app load performance.
Summary:
This diff introduces a change to the VulkanStagingDevice class, to use multiple buffers for staging data instead of a single buffer.
It also automatically free unused blocks of staging buffer to reduce GPU memory footprint.

This is done to improve app load time performance by eliminating waits on already issued copy commands. The change also includes updates to the VulkanCommandQueue class to call a new method in the VulkanStagingDevice class to merge regions and free buffers.

Reviewed By: corporateshark

Differential Revision: D55540178

fbshipit-source-id: 0cc86961925c68cecd9add573fe788e7f2bebec6