Storing VulkanImage object in VulkanTexture instead of unique ptr reference.
Summary: This diff makes changes to the IGL Vulkan codebase to store a VulkanImage object in a VulkanTexture instead of using a unique_ptr reference. This change is made to improve memory management. The changes include modifying the VulkanImage class to store the image and image view objects as members instead of using a unique_ptr reference, modifying the VulkanContext class to create VulkanImage objects directly instead of using a unique_ptr reference, modifying the VulkanSwapchain class to store VulkanImage objects directly instead of using a unique_ptr reference, and modifying the IGLSurfaceTextureVulkan class to store VulkanImage objects directly instead of using a unique_ptr reference.

Reviewed By: corporateshark

Differential Revision: D55622632

fbshipit-source-id: 3b9c00255a9ce20da3a779cd29dbcd5771441775