Fix to vulkan staging buffer failures happening due to memory greater than maxStagingBufferSize_ being requested.
Summary: This diff fixes a bug in the Vulkan staging buffer that occurs when a memory request greater than the maximum staging buffer size is made. The code changes in the VulkanStagingDevice.cpp file include adding a check to see if the requested size is available or if contiguous memory is not requested.

Reviewed By: corporateshark

Differential Revision: D56159665

fbshipit-source-id: c7f7eaf5125c42019d8306dfa1f1b169a87ab266