Adding vmaPreferredLargeHeapBlockSize to VulkanContextConfig to explicitly control VmaAllocatorCreateInfo::preferredLargeHeapBlockSize.
Summary: This diff adds a new field to the `VulkanContextConfig` struct called `vmaPreferredLargeHeapBlockSize` which allows users to explicitly control the `preferredLargeHeapBlockSize` parameter of the `VmaAllocatorCreateInfo` struct used by the Vulkan Memory Allocator (VMA) library. This parameter controls the size of the largest block of memory that VMA will allocate from the system heap. By default, VMA will use a large default value of 256 MB, but this can be reduced or increased depending on the specific needs of the application. The `vmaPreferredLargeHeapBlockSize` field is only relevant if VMA is used for memory allocation, and passing 0 will prompt VMA to use the default value.

Reviewed By: corporateshark

Differential Revision: D56255557

fbshipit-source-id: 6318d211b111b7d480f0a112047cb5ae8c9e3207