Add blit function to RenderCommandEncoder
Summary:
Add blitColorImage() to RenderCommandEncoder.h for PanelApp Render Shortcut.
Blit is used together with imageData().
Currently blit is only used for color image so this function did not handle stencil/depth and layers.

It is a retreat from D54715295. Thought twice, the blitColorImage() function could just live on top of igl library because it is not that generic.

Reviewed By: corporateshark, mmaurer

Differential Revision: D55523190

fbshipit-source-id: 1e81cce8d683dfb772498e29481f093a5716904e