Re-enable enumeration of extensions on Android
Summary: Enumeration of extensions has been disabled in Android due to an older bug in validation layers - running in debug mode causes extension enumeration to crash. Disabling the check has led to configuration guards for specific platforms rather than addressing the root cause. This change re-enables enumeration of extensions to support more varied platform configurations (e.g. emulators on mac) and also in an effort to root cause the bug in the android validation layer (if it isn't fixed yet.)

Reviewed By: corporateshark, mmaurer

Differential Revision: D56281969

fbshipit-source-id: cad5ae098d7146792121a4eaa5878acff697dc52