Check whether OpenGl Object IDs are valid after calling the ID generation APIs
Summary:
Sometimes the OpenGlContext might be accidentally deleted by other part of the same application. The ID generation will fail in this case. Before this change, the rest of the function would continue to run in release build mode. Sometimes it doesn't even report any failure even though it fails completely under the hood.

This change makes sure to check the IDs and report errors immediately.

Reviewed By: EricGriffith, tgoulart

Differential Revision: D56645569

fbshipit-source-id: 59084f11fd00fdef2cd5005c0655a94c833a70e7