Removing redundant localData_ buffer allocation and memcpy when Vulkan Buffer memory is mapped to CPU.
Summary: This diff removes the redundant localData_ buffer allocation and memcpy when Vulkan Buffer memory is mapped to CPU. The changes made in the Buffer.cpp file include removing the localData_ buffer allocation and memcpy when isRingBuffer_ is true. This change improves performance by reducing unnecessary memory allocation and copying.

Reviewed By: syeh1, corporateshark

Differential Revision: D57500236

fbshipit-source-id: dcca9e800d0af668c6e65d8dce7d009b2703afe9