igl | Fix Intel Mac multisample textures
Summary:
We are seeing an error on Intel Mac after adding multisample support in environments of simulator, that suggests we can't enforce storage mode as we are doing now.

We are explicitly setting the mode to Private, and we need to avoid this override.

Error message shows on logview:
-[MTLTextureDescriptorInternal validateWithDevice:]:1357: failed assertion `Texture Descriptor Validation
MTLTextureDescriptor: Multisample textures cannot be allocated with MTLStorageModeShared or MTLStorageModeManaged on this device.
'

Reviewed By: pixelperfect3

Differential Revision: D57357905

fbshipit-source-id: ded7d8391fd43d41ec2b0deffceb284589dce172