Blend modes support
Summary: This diff introduces support for blend modes in the AtlasSandbox app. It includes changes to the XrApp.h and XrApp.cpp files to add support for the XR_FB_composition_layer_alpha_blend extension, as well as changes to the QuadLayerParams.h and GPUStressSession.cpp files to include a new LayerBlendMode enum and support for setting the layer alpha blend mode. The Factory.cpp file includes changes to set the blend mode based on a parameter passed in from the app.

Reviewed By: pixelperfect3, corporateshark

Differential Revision: D57757446

fbshipit-source-id: c9257e618c177c8fdd3509113cd9d9dd97ce9b21