Fix for missing emplace_back
Summary:
The diff titled "[Stresstest] Fix for missing push_back" contains code changes in the file "GPUStressSession.cpp" in the "igl/public/shell/renderSessions" directory. The changes involve fixing a missing push_back operation in the code. The code change is as follows:

```
params.layerAlphaBlend_.emplace_back(igl::shell::LayerBlendMode::AlphaBlend);
```

This change is necessary to fix a missing push_back operation that was causing a runtime error. The changes were made to ensure that the code runs correctly and does not cause any errors.

Reviewed By: EricGriffith, nlutsenko

Differential Revision: D57929266

fbshipit-source-id: d3ca9d1b703ebe6d1c89897d5b7628a72db419c5