Compute max uniform length by taking the max of GL_ACTIVE_UNIFORM_MAX_LENGTH with the max of each GL_UNIFORM_NAME_LENGTH
Summary:
Compute max uniform length by querying GL_ACTIVE_UNIFORM_MAX_LENGTH, and then taking the max of that with every GL_UNIFORM_NAME_LENGTH of each of the uniforms.

This is needed because we observed that OpenGL drivers are sometimes unreliable with these values:

1. Android devices with old Mali GPUs (e.g. Mali-T860MP2) sometimes incorrectly return 0 for  GL_ACTIVE_UNIFORM_MAX_LENGTH

2. When running macOS unit tests, sometimes GL_UNIFORM_NAME_LENGTH always return 0

So the safe thing to do here is to take the max of the two.

Reviewed By: dmannemela, syeh1

Differential Revision: D57974011

fbshipit-source-id: 44cc81e6bcad7619eb17f321a402f92de694b1fa