Remove incorrect collision check
Summary:
This check doesn't make sense, consider the following sequence of calls:
```
// drawing the first object
encoder.bindBuffer (0, bufferA)
encoder.bindVertexBuffer(1, bufferB)
encoder.draw(...)

// drawing a different object
encoder.bindBuffer (1, bufferC)
encoder.bindVertexBuffer(0, bufferD)
encoder.draw(...)
```
This is perfectly valid (and will work correctly), but this check here determines this to be a collision.

There is an effect 778223544407551 which issues commands using this pattern and which crashes on IOS due to this check.

Reviewed By: EricGriffith, KyleFung, corporateshark

Differential Revision: D58029138

fbshipit-source-id: 7f266cc965f9e63f6414d527ecdeea5aa063e1a0