Add support to read color attachment to CPU memory whose expected bytesPerRow is larger than (width * componentsPerPixel + alignment)
Summary: In the copyBytesColorAttachment function, we allow API users to pass in bytesPerRow to specify the row size in their CPU buffer. However, the bytesPerRow currently is only used for computing the alignment. Sometimes, end user might pass in a buffer whose width is much larger than the texture itself (e.g. I want to read the 1080 width into my 1440 width buffer). This is a very useful feature to add in the IGL OpenGL and it can be supported by setting the GL_PACK_ROW_LENGTH to the expected value as shown in this diff.

Reviewed By: EricGriffith, francoiscoulombe

Differential Revision: D58214890

fbshipit-source-id: 690d563bd7fbb8efa818e540563f7c19dabe9f30