Use ImageWriterSTB on Android (#140)
Summary:
Pull Request resolved: https://github.com/facebook/igl/pull/140

Writing images in IGL on Android platform was stubbed with "not implemented" stub, which asserted.

It should be just fine using the regular `ImageWriterSTB`, which writes to a file system - arguably, writing into the app's resources doesn't make much practical sense, and we don't have to be symmetrical with the reader, which *does* read from the resources (which the stub appeared to assume).

In RN/Granite, we want this functionality to be available for:
* running screenshot tests on device
* caching images on disk, which are created from pixel data
* debugging (which has been indeed very useful in my case, when debugging OpenGL ES drawing)

Reviewed By: christophpurrer

Differential Revision: D58747911

fbshipit-source-id: bb129c9f6e7e83c3fe5c167303ab4c429493a1a8