Copying texture data directly when upload is performed on a texture with mapped memory.
Summary: The change is related to copying texture data directly when upload is performed on a texture with mapped memory. The code change involves checking if the texture has mapped memory and if so, copying the data directly to the mapped memory instead of using the staging device. This change is expected to improve performance by reducing the number of memory transfers.

Reviewed By: syeh1, corporateshark

Differential Revision: D58272455

fbshipit-source-id: 90925cae0ff8a5ee2307de20e2331d8f02c6f38d