Remove OpenGL dependency from Vulkan tests
Summary:
# Problem:
We link 2 versions of swiftshader (old one with OpenGl support, new one with Vulkan). The first is linked inside IGL's binaries, the second is used as dynamic library. It creates difficulties for TSAN, even if Vulkan doesn't use OpenGL's swiftshader.

# Solution:
Don't link OpenGL to `iglVulkanTests` for Linux where we execute TSAN. Later it can be propagated to other platforms.

Reviewed By: rudybear, corporateshark

Differential Revision: D59045064

fbshipit-source-id: 988d36e58e4024ec3d6a3155c2deb71aca76ccf8